Tighten the FiberIO contract its plugins were already keeping (#816)

+++ dascore/io/index/planned.py
@@ -129,7 +129,7 @@ def _coord_record_from_row(
     if step is not None:
         # lo, hi, and step always share a time kind (or are all floats), but
         # ty unions the branch types and rejects the mixed combinations.
-        length = int(round((hi - lo) / step)) + 1  # ty: ignore[unsupported-operator]
+        length = round((hi - lo) / step) + 1  # ty: ignore[unsupported-operator]
     key = row.get(f"_{name}_def_key")
     fingerprint = None
     if isinstance(key, str) and key.startswith("fp:"):
